requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
numpy==1.16.2
python-dotenv==0.10.1
requests==2.21.0
Shapely==1.6.4.post2
six==1.12.0
slackclient==1.3.1
urllib3==1.24.1
websocket-client==0.54.0
